requirements.txt
Django==5.0.2
psycopg2-binary==2.9.10
dj-database-url==2.1.0
Pillow==11.3.0
django-extensions==3.2.3
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-filter==23.5
reportlab==4.0.8
openpyxl==3.1.2
python-dateutil==2.8.2
qrcode==7.4.2
pyotp==2.9.0
cryptography==42.0.8
bleach==6.1.0

# تحسينات الأداء
django-redis==5.4.0
django-compressor==4.4
gunicorn==21.2.0
whitenoise==6.6.0
django-debug-toolbar==4.2.0
django-cachalot==2.6.1